requirements.txt
python==3.7.12
numpy==1.21.6
scikit-learn==1.0.2
pytorch==1.11.0
torchvision==0.12
joblib==1.1.0
matplotlib==3.5.1
seaborn==0.12.0
pandas==1.3.5
torchmetrics==0.8.1
scipy==1.7.3
statsmodels==0.13.2
pytest==7.1.2
pytest-xdist==3.3.1
datasets==2.12.0
transformers==4.16.2
hypothesis==6.46.1
progressbar==2.5
